project/requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
django-admin-sortable2>=1.0
django-tinymce>=3.5.0
environs>=10.0.0
requests>=2.31.0
python-slugify>=8.0.0

